requirements.txt
appdirs==1.4.4
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
black==20.8b1
click==7.1.2
colorama==0.4.4
Django==3.1.7
django-extensions==3.1.1
django-filter==2.4.0
djangorestframework==3.12.2
dnspython==2.1.0
email-validator==1.1.2
Faker==6.4.1
flake8==3.8.4
idna==3.1
importlib-metadata==3.4.0
iniconfig==1.1.1
Markdown==3.3.3
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pydantic==1.7.3
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
zipp==3.4.0
